Index as param is size_t

+++ generator.h
@@ -18,10 +18,10 @@ class Generator {
   std::ifstream &in_file;
  public:
   Generator(const std::vector<Compressor*> &Compressors, std::ifstream &in_file);
-  void run(int index);
+  void run(size_t index);
   std::vector<std::queue<std::string>*> get_outputs();
   ~Generator();
-  void operator()(int index);
+  void operator()(size_t index);
  private:
   size_t calc_offset(size_t size_block, size_t index, size_t run_number);
 };
